feat: verify if user exists on delete

+++ src/use-cases/user-use-case.ts
@@ -36,6 +36,8 @@ export class UserUserCase {
   }
 
   async remove(id: string): Promise<void> {
+    const verifyIfUserExists = await this.userRepository.findByID(id);
+    if (!verifyIfUserExists) throw new Error("User Not Found");
     await this.userRepository.remove(id);
   }
 }
